# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mapa de Emergencia y Rescate — API",
    "version": "1.0.0",
    "description": "API pública/admin del mapa de emergencia. Documentación generada automáticamente desde los bloques @swagger de cada route."
  },
...
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "MissingPerson": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "age": {
            "type": "integer",
            "nullable": true
          },
          "description": {
            "type": "string"
          },
          "lastSeen": {
            "type": "string"
          },
          "contact": {
            "type": "string"
          },
          "photoUrl": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "found"
            ]
          },
          "resolutionNote": {
            "type": "string",
            "nullable": true
          },
          "resolutionPhotoUrl": {
            "type": "string",
            "nullable": true
          },
          "resolvedAt": {
            "type": "integer",
            "nullable": true
          },
          "createdAt": {
            "type": "integer",
            "description": "epoch-ms"
          }
        }
      },
...
      "HospitalPatient": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "hospitalId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "age": {
            "type": "integer",
            "nullable": true
          },
          "condition": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "contact": {
            "type": "string"
          },
          "admittedAt": {
            "type": "integer"
          },
          "updatedAt": {
            "type": "integer"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/missing": {
      "get": {
        "tags": [
          "missing"
        ],
        "summary": "Lista paginada de personas desaparecidas o localizadas",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "found",
                "all"
              ],
              "default": "active"
            },
            "description": "Filtro de estado. Por defecto solo activas."
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            },
            "description": "Término de búsqueda por nombre."
          }
        ],
        "responses": {
          "200": {
            "description": "Página de personas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "people": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MissingPerson"
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "totalCapped": {
                      "type": "boolean"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "pageSize": {
                      "type": "integer"
                    },
                    "totalPages": {
                      "type": "integer"
                    },
                    "persistent": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "missing"
        ],
        "summary": "Crea un reporte de persona desaparecida o localizada",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "age": {
                    "type": "integer",
                    "nullable": true
                  },
                  "description": {
                    "type": "string"
                  },
                  "lastSeen": {
                    "type": "string"
                  },
                  "contact": {
                    "type": "string"
                  },
                  "photo": {
                    "type": "string",
                    "nullable": true,
                    "description": "Data URL de imagen JPG, PNG o WebP"
                  },
                  "reportType": {
                    "type": "string",
                    "enum": [
                      "missing",
                      "found"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reporte creado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "person": {
                      "$ref": "#/components/schemas/MissingPerson"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Foto demasiado grande",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Límite de tasa excedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "No se pudo guardar el reporte",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/patients/search": {
      "get": {
        "tags": [
          "hospitals"
        ],
        "summary": "Busca pacientes hospitalizados por nombre u otros datos",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Texto de búsqueda"
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            },
            "description": "Máximo de resultados (se acota entre 1 y 200)"
          }
        ],
        "responses": {
          "200": {
            "description": "Resultados de la búsqueda de pacientes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "results": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/HospitalPatient"
                      }
                    },
                    "query": {
                      "type": "string"
                    },
                    "hasMore": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/reports/{id}": {
      "delete": {
        "tags": [
          "reports"
        ],
        "summary": "Marca un reporte como atendido (lo elimina). Solo administradores.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reporte eliminado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Falta el id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado (se requiere administrador)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Reporte no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Demasiadas solicitudes (rate limit)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}